+++ tests/testthat/test-time-varying.R
@@ -85,7 +85,7 @@ test_that("AR1 time-varying works", {
                 time_varying = ~ -1 + depth_scaled,
                 data = pcod_2011, spatial="off",
                 spatiotemporal = "off",
-                family = tweedie())
+                family = tweedie(), priors = sdmTMBpriors(sigma_V = gamma_cv(0.2, 0.5)))
   s <- tidy(fit, "ran_pars")
   expect_equal(dim(s), c(2L, 5L))
   expect_equal(s$term, c("phi","tweedie_p"))
@@ -96,7 +96,7 @@ test_that("AR1 time-varying works", {
                 data = pcod_2011, spatial="off",
                 time_varying_type = "ar1",
                 spatiotemporal = "off",
-                family = tweedie())
+                family = tweedie(), priors = sdmTMBpriors(sigma_V = gamma_cv(0.2, 0.5)))
   s <- tidy(fit, "ran_pars")
   expect_equal(dim(s), c(3L, 5L))
   expect_equal(s$term, c("phi", "rho_time", "tweedie_p"))
@@ -122,24 +122,24 @@ test_that("AR1 time-varying works", {
   expect_equal(s$term, c("phi","rho_time","rho_time","tweedie_p"))
 
   # test that tidy works with delta
-  fit <- sdmTMB(density ~ 1, time = "year",
-                time_varying = ~ -1+depth_scaled + I(depth_scaled^2),
+  fit <- sdmTMB(density ~ depth_scaled, time = "year",
+                time_varying = ~ -1+I(depth_scaled^2),
                 time_varying_type = "ar1",
                 data = pcod_2011, spatial="off",
                 spatiotemporal = "off",
-                family = delta_gamma())
+                family = delta_gamma(), priors = sdmTMBpriors(sigma_V = gamma_cv(0.2, 0.5)))
   s <- tidy(fit, "ran_pars")
-  expect_equal(dim(s), c(2L, 6L))
-  expect_equal(s$term, c("rho_time","rho_time"))
+  expect_equal(dim(s), c(1L, 6L))
+  expect_equal(s$term, c("rho_time"))
 
   s <- tidy(fit, "ran_pars", model = 1)
-  expect_equal(dim(s), c(2L, 6L))
-  expect_equal(s$term, c("rho_time","rho_time"))
-  expect_equal(s$estimate, c(1.0, 0.9582931))
+  expect_equal(dim(s), c(1L, 6L))
+  expect_equal(s$term, c("rho_time"))
+  expect_equal(s$estimate, c(0.888583727), tolerance = 0.001)
 
   s <- tidy(fit, "ran_pars", model = 2)
-  expect_equal(s$term, c("phi", "rho_time","rho_time"))
-  expect_equal(s$estimate, c(0.652, 1.0, 0.868), tolerance = 0.001)
+  expect_equal(s$term, c("phi", "rho_time"))
+  expect_equal(s$estimate, c(0.65029, 0.70935), tolerance = 0.001)
 
 })